Arduino Nano atmega328p
Power of sight is considered as the most important of all senses. Visually challenged people are often dependent on others for their daily duties. Over time several advancements in technology has helped in increasing entertainment and comfort for the visually challenged. “Smart shoe" can assist the visually challenged on their daily routine and can act as a comfortable and safe companion on their journey.